services/core/pyproject.toml
[tool.poetry]
name = "core"
version = "1.1.3"
description = "Core PLO (Pot Limit Omaha) solver and equity calculation library"
authors = ["PLOSolver Team <hello@plosolver.com>"]
maintainers = ["PLOSolver Team <hello@plosolver.com>"]
readme = "README.md"
license = "MIT"
keywords = ["poker", "plo", "solver", "equity", "gto"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
    "Topic :: Scientific/Engineering :: Mathematics",
]
include = ["README.md", "LICENSE.txt"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.0,<2.0.0"
flask_jwt_extended = "^4.6.0,<4.7.0"
flask_socketio = "^5.5.1"
scikit-learn = "^1.3.0,<2.0.0"
joblib = "^1.3.0,<2.0.0"
treys = "^0.1.8,<1.0.0"
deuces = "0.2"
celery = "^5.3.0,<5.3.20"
requests = "^2.32.4,<2.32.10"
pika = "^1.3.0,<1.4.1"
aio-pika = "^9.5.0,<9.8.1"
redis = "^4.6.0,<4.7.1"
# Database dependencies removed - core package no longer includes database functionality
poetry-core = "^2.1.3"
protobuf = "^6.32.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
flake8-pyproject = "^1.2.0"
mypy = "^1.0.0"
# Flask database extensions removed - core package no longer includes database functionality
pre-commit = "^3.6.0"
isort = "^5.13.0"
bandit = "^1.7.5"
pylint = "^3.3.8"
ruff = "^0.6.9"

[tool.poetry.group.test.dependencies]
hatchling = "^1.20.0"
pytest = "^7.0.0"
# Flask database extensions removed - core package no longer includes database functionality

[tool.poetry.urls]
Homepage = "https://github.com/your-org/PLOSolver"
Repository = "https://github.com/your-org/PLOSolver"
Documentation = "https://github.com/your-org/PLOSolver#readme"
Issues = "https://github.com/your-org/PLOSolver/issues"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]
ignore = ["E203"]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["core"]

[tool.ruff.format]
quote-style = "double"

[tool.flake8]
max-line-length = 120
# E203: Whitespace before ':' (conflicts with Black formatting)
# W503: Line break before binary operator (conflicts with Black formatting)
ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
mypy_path = "."
explicit_package_bases = true

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101", "B105", "B110", "B112", "B311", "B324"]
